requirements.txt
fastapi==0.115.0
lightrag-hku>=1.4.9.8
networkx>=3.4
numpy>=2.2
ollama>=0.5.0
orjson>=3.10
pillow>=11.0.0
pydantic>=2.10
pydantic-settings>=2.6
PyMuPDF>=1.25.0
PyYAML>=6.0.2
pytesseract>=0.3.13
python-dotenv>=1.0.1
sentence-transformers==5.3.0
streamlit>=1.39.0
torch==2.10.0
tqdm>=4.67.0
transformers==4.57.6
uvicorn[standard]>=0.32.0
websockets>=13.1
